Added ability to make custom database call in Login module
